Move ALL_DECKS_ID constant to Decks extensions file
Extra step in removing DeckSpinnerSelection
