+++ anki-common/src/main/java/com/ichi2/anki/common/Constants.kt
@@ -1,0 +1,10 @@
+/*
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+package com.ichi2.anki.common
+
+/**
+ * [com.ichi2.anki.libanki.DeckId] constant to represent "All decks" in screens that need it(ex. browser).
+ */
+const val ALL_DECKS_ID = 0L
